requirements.txt
logfire[fastapi,asyncpg]
fastapi
uvicorn[standard]
pydantic-ai[google]
pydantic
python-multipart
python-dotenv
asyncpg